Правки ошибки Escape
